# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Resource Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentFolderId?: number | null;
  resourceCount?: number;
  createdAt: string;
}

export interface FolderInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  parentFolderId?: number | null;
}

export interface FolderUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  parentFolderId?: number | null;
}

export type ResourceType = typeof ResourceType[keyof typeof ResourceType];


export const ResourceType = {
  link: 'link',
  file: 'file',
  video: 'video',
} as const;

export interface Resource {
  id: number;
  name: string;
  type: ResourceType;
  url: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export type ResourceInputType = typeof ResourceInputType[keyof typeof ResourceInputType];


export const ResourceInputType = {
  link: 'link',
  file: 'file',
  video: 'video',
} as const;

export interface ResourceInput {
  /** @minLength 1 */
  name: string;
  type: ResourceInputType;
  url: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}

export type ResourceUpdateType = typeof ResourceUpdateType[keyof typeof ResourceUpdateType];


export const ResourceUpdateType = {
  link: 'link',
  file: 'file',
  video: 'video',
} as const;

export interface ResourceUpdate {
  /** @minLength 1 */
  name?: string;
  type?: ResourceUpdateType;
  url?: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}

export type StatsByType = {
  link: number;
  file: number;
  video: number;
};

export interface Stats {
  totalFolders: number;
  totalResources: number;
  byType: StatsByType;
}

export type ListResourcesParams = {
folderId?: number;
type?: ListResourcesType;
search?: string;
};

export type ListResourcesType = typeof ListResourcesType[keyof typeof ListResourcesType];


export const ListResourcesType = {
  link: 'link',
  file: 'file',
  video: 'video',
} as const;

export type ListRecentResourcesParams = {
limit?: number;
};