# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Resource Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from './resourceType';

export interface Resource {
  id: number;
  name: string;
  type: ResourceType;
  url: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  createdAt: Date;
  updatedAt?: Date;
}