# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Resource Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceInputType } from './resourceInputType';

export interface ResourceInput {
  /** @minLength 1 */
  name: string;
  type: ResourceInputType;
  url: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}