# lib/api-zod/src/generated/types/resourceUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Resource Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceUpdateType } from './resourceUpdateType';

export interface ResourceUpdate {
  /** @minLength 1 */
  name?: string;
  type?: ResourceUpdateType;
  url?: string;
  /** @nullable */
  folderId?: number | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
}